requirements.txt
streamlit
experta
requests
huggingface_hub
python-dotenv